requirements.txt
# Core dependencies
langgraph>=0.0.3
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.5.0
langchain_ollama>=0.0.1
langgraph_swarm>=0.0.1

# Document processing
pypdf2>=3.0.0
python-docx>=0.8.11
spacy>=3.7.0

# LLM and API clients
requests>=2.31.0
ollama>=0.1.0
openai>=1.3.0
langchain_google_genai>=0.0.1
google-generativeai>=0.3.0
langchain_core>=0.1.0

# Testing and development
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.9.1
isort>=5.12.0
mypy>=1.5.1

# Utilities
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0
bs4>=0.0.1
duckduckgo_search>=4.1.0
clean-text>=0.6.0
qdrant_client>=1.7.0
uuid>=1.30
beautifulsoup4>=4.12.0
tavily-python